backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic>=2.0.0
pandas==2.2.3
python-multipart==0.0.20
openpyxl==3.1.5
numpy>=1.24.0
anthropic>=0.84.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-pptx>=1.0.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
cryptography>=41.0.0
httpx>=0.26.0
stripe>=8.0.0
pip-audit>=2.7.0
alembic>=1.13.0
rq>=1.16.0
redis>=5.0.0
